refactor: scoringService.ts を scoring/ モジュールに分解 #6
